+++ frontend/src/pages/index.js
@@ -9,7 +9,7 @@ import ReCAPTCHA from "react-google-recaptcha";
 import styles from "@/styles/Home.module.css";
 import Input from "@/components/ui/Input/Input";
 import Button from "@/components/ui/Button/Button";
-import Logo from "@/components/ui/Logo/Logo";
+import Logo from "@/assets/logo/Logo";
 import SnsButtons from "@/components/ui/SnsButtons/SnsButtons";
 import TitleWeb from "@/assets/title/TitleWeb";
 import TitleMobile from "@/assets/title/TitleMobile";
